requirements.txt

### requirements.txt
text
# Core
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.41.0
accelerate>=0.30.0

# Data & audio
librosa>=0.10.1
soundfile>=0.12.1
pandas>=2.1.0
numpy>=1.26.0

# Training & evaluation
scikit-learn>=1.3.0
tqdm>=4.66.0
PyYAML>=6.0.1
matplotlib>=3.8.0
